+++ packages/state/src/api.ts
@@ -7,13 +7,13 @@ import {
   StoreApiDefinition,
 } from '~/types';
 import { onCleanup } from 'solid-js';
-import { ApiDefinition } from '~/apiDefinition';
+import { ApiDefinition } from '~/api-definition';
 import { Container } from '~/container';
 import { ResolvedPluginContext } from '~/resolved-plugin-context';
+import { StateBuilderError } from '~/error';
 
 export const $CREATOR = Symbol('store-creator-api'),
   $PLUGIN = Symbol('store-plugin');
-
 /**
  * A factory function that creates a store API definition creator.
  *
@@ -44,8 +44,8 @@ function checkDependencies(
 
   dependencies.forEach((dependency) => {
     if (!resolvedPlugins.includes(dependency)) {
-      throw new Error(
-        `[statebuilder] The dependency '${dependency}' of plugin '${meta.name}' is missing`,
+      throw new StateBuilderError(
+        `The dependency '${dependency}' of plugin '${meta.name}' is missing`,
         { cause: { resolvedDependencies: resolvedPlugins, plugin } },
       );
     }
@@ -63,10 +63,7 @@ function checkDependencies(
 export function resolve<
   TDefinition extends StoreApiDefinition<GenericStoreApi, Record<string, any>>,
 >(definition: TDefinition, container?: Container) {
-  const api = definition[$CREATOR];
-
-  const { factory, plugins } = api;
-
+  const { factory, plugins } = definition[$CREATOR];
   const resolvedPlugins: string[] = [],
     pluginContext = new ResolvedPluginContext(container, plugins),
     resolvedStore = factory();
@@ -92,8 +89,21 @@ export function resolve<
     resolvedPlugins.push(extensionCreator.name);
   }
 
-  pluginContext.runInitSubscriptions(resolvedStore);
-  onCleanup(() => pluginContext.runDestroySubscriptions(resolvedStore));
+  if (!!container) {
+    pluginContext.hooks.onDestroy(() => container.remove(definition));
+  }
+
+  for (const listener of pluginContext.initSubscriptions) {
+    listener(resolvedStore);
+    pluginContext.initSubscriptions.delete(listener);
+  }
+
+  onCleanup(() => {
+    for (const listener of pluginContext.destroySubscriptions) {
+      listener(resolvedStore);
+      pluginContext.destroySubscriptions.delete(listener);
+    }
+  });
 
   return resolvedStore;
 }